debug: unit tests reorganisation
